Fixed issues with exporting variables to submenu.sh
This menu is called multiple times in a number of scripts, so the exported variables for display were being re-displayed if
the next call didn't used as many variables. Fixed by adding the following before the variables are exported for each
display.

for each in text{1..6};do unset $each;done